requirements.txt
asgiref
attrs
autopep8
certifi
charset-normalizer
crispy-bootstrap5
distlib
dj-database-url
Django
django-cors-headers
django-crispy-forms
django-environ
django-extensions
django-filter
django-silk
djangorestframework
djangorestframework_simplejwt
drf-spectacular
Faker
filelock
gprof2dot
idna
inflection
jsonschema
jsonschema-specifications
Markdown
MouseInfo
packaging
pillow
pipenv
platformdirs
psycopg2-binary
PyAutoGUI
pycodestyle
PyGetWindow
PyJWT
PyMsgBox
pyperclip
PyRect
PyScreeze
python-dateutil
python-decouple
python-dotenv
python-environ
python3-xlib
pytweening
PyYAML
referencing
requests
rpds-py
setuptools
six
sqlparse
stripe
typing_extensions
tzdata
uritemplate
urllib3
virtualenv
whitenoise
